IOS-297 Parse problem doc messages per API docs
Also refactor methods showing alerts.
